maj synchro

+++ client/include/Definitions.hh
@@ -27,7 +27,7 @@ namespace client
 typedef std::pair<int16_t, int16_t>	pos_t;
 typedef std::pair<int16_t, int16_t>	vec_t;
 typedef uint32_t			tick;
-typedef uint16_t			ide_t;
+typedef uint32_t			ide_t;
 typedef uint16_t			typeide_t;
 typedef int16_t				UIevent_t;
